SmartEnemy

+++ src/GameObject.java
@@ -2,13 +2,13 @@ import java.awt.*;
 
 public abstract class GameObject {
 
-    protected int x,y;
+    protected float x,y;
     protected ID id;
 
 
-    protected int velX, velY;
+    protected float velX, velY;
 
-    public GameObject(int x, int y, ID id){
+    public GameObject(float x, float y, ID id){
         this.x = x;
         this.y = y;
         this.id = id;
@@ -26,8 +26,8 @@ public abstract class GameObject {
     public void setY(int y){
         this.y = y;
     }
-    public int getX(){ return x; }
-    public int getY(){
+    public float getX(){ return x; }
+    public float getY(){
         return y;
     }
     public void setId(ID id){
@@ -39,13 +39,13 @@ public abstract class GameObject {
     public void setVelX(int velX) {
         this.velX = velX;
     }
-    public int getVelX() {
+    public float getVelX() {
         return velX;
     }
     public void setVelY(int velY) {
         this.velY = velY;
     }
-    public int getVelY() {
+    public float getVelY() {
         return velY;
     }
